task #21 Class rename
