+++ vaadin-spreadsheet/src/test/java/com/vaadin/addon/spreadsheet/test/AutofitRowTest.java
@@ -1,6 +1,6 @@
 package com.vaadin.addon.spreadsheet.test;
 
-import static com.vaadin.addon.spreadsheet.test.testutil.TextLineHeightChecker.assertThatCellHeightIsAcceptable;
+import static com.vaadin.addon.spreadsheet.test.testutil.CellHeightChecker.assertThatCellHeightIsAcceptable;
 import java.util.List;
 
 import org.junit.Test;
@@ -8,7 +8,10 @@ import org.openqa.selenium.remote.DesiredCapabilities;
 
 import com.vaadin.addon.spreadsheet.elements.SheetCellElement;
 import com.vaadin.addon.spreadsheet.elements.SpreadsheetElement;
+import com.vaadin.testbench.annotations.RunLocally;
+import com.vaadin.testbench.parallel.Browser;
 
+@RunLocally(Browser.PHANTOMJS)
 public class AutofitRowTest extends AbstractSpreadsheetTestCase {
 
     public static final int HUGE_FONT_SIZE = 48;
